Added java.command package that contains Allocator, Processor, Response classes. And also made some minor edits
