+++ src/main/java/duke/command/Processor.java
@@ -1,0 +1,11 @@
+package duke.command;
+
+public class Processor {
+
+    /* Split input into:
+       keyword OR
+       keyword + num OR
+       keyword + description + /by or /at + time
+     */
+
+}
