+++ src/main/java/duke/task/Deadline.java
@@ -2,7 +2,7 @@ package duke.task;
 
 public class Deadline extends Task {
 
-    protected String by;
+    private String by;
 
     public Deadline(String description, String by) {
         super(description);
