+++ src/main/java/duke/task/Event.java
@@ -2,7 +2,7 @@ package duke.task;
 
 public class Event extends Task {
 
-    protected String at;
+    private String at;
 
     public Event(String description, String at) {
         super(description);
