+++ text-ui-test/EXPECTED.TXT
@@ -1,56 +1,56 @@
-__________________________________________________
+___________________________________________________________________________
 Hello! I'm Duke
 What can I do for you?
-__________________________________________________
-__________________________________________________
+___________________________________________________________________________
+___________________________________________________________________________
 Got it. I've added this task:
 [T][ ] read book
 Now you have 1 tasks in the list.
-__________________________________________________
-__________________________________________________
+___________________________________________________________________________
+___________________________________________________________________________
 Got it. I've added this task:
 [D][ ] return book (by: Sunday)
 Now you have 2 tasks in the list.
-__________________________________________________
-__________________________________________________
+___________________________________________________________________________
+___________________________________________________________________________
 Got it. I've added this task:
 [E][ ] project meeting (at: Mon 2-4pm)
 Now you have 3 tasks in the list.
-__________________________________________________
-__________________________________________________
+___________________________________________________________________________
+___________________________________________________________________________
 Got it. I've added this task:
 [D][ ] do homework (by: no idea :-p)
 Now you have 4 tasks in the list.
-__________________________________________________
-__________________________________________________
+___________________________________________________________________________
+___________________________________________________________________________
 Here are the tasks in your list:
 1.[T][ ] read book
 2.[D][ ] return book (by: Sunday)
 3.[E][ ] project meeting (at: Mon 2-4pm)
 4.[D][ ] do homework (by: no idea :-p)
-__________________________________________________
-__________________________________________________
+___________________________________________________________________________
+___________________________________________________________________________
 Nice! I've marked this task as done:
 [D][X] do homework (by: no idea :-p)
-__________________________________________________
-__________________________________________________
+___________________________________________________________________________
+___________________________________________________________________________
 Here are the tasks in your list:
 1.[T][ ] read book
 2.[D][ ] return book (by: Sunday)
 3.[E][ ] project meeting (at: Mon 2-4pm)
 4.[D][X] do homework (by: no idea :-p)
-__________________________________________________
-__________________________________________________
+___________________________________________________________________________
+___________________________________________________________________________
 OK, I've marked this task as not done yet:
 [D][ ] do homework (by: no idea :-p)
-__________________________________________________
-__________________________________________________
+___________________________________________________________________________
+___________________________________________________________________________
 Here are the tasks in your list:
 1.[T][ ] read book
 2.[D][ ] return book (by: Sunday)
 3.[E][ ] project meeting (at: Mon 2-4pm)
 4.[D][ ] do homework (by: no idea :-p)
-__________________________________________________
-__________________________________________________
+___________________________________________________________________________
+___________________________________________________________________________
 Bye. Hope to see you again soon!
-__________________________________________________
+___________________________________________________________________________
